Items State action confirmations

+++ src/app/components/batch/actions/actions-item-state/delete-pub-items-form/delete-pub-items-form.component.ts
@@ -1,8 +1,9 @@
 import { CommonModule } from '@angular/common';
-import { Component } from '@angular/core';
+import { Component, inject } from '@angular/core';
 import { Router } from '@angular/router';
 
 import { BatchService } from 'src/app/components/batch/services/batch.service';
+import { MessageService } from 'src/app/shared/services/message.service';
 import type { DeletePubItemsParams } from 'src/app/components/batch/interfaces/batch-params';
 
 @Component({
@@ -13,11 +14,10 @@ import type { DeletePubItemsParams } from 'src/app/components/batch/interfaces/b
   ],
   templateUrl: './delete-pub-items-form.component.html'
 })
-export class DeletePubItemsFormComponent { 
-
-  constructor(
-    private router: Router,
-    private batchSvc: BatchService) {}
+export class DeletePubItemsFormComponent {
+  batchSvc = inject(BatchService);
+  msgSvc = inject(MessageService);
+  router = inject(Router);
 
   get deletePubItemsParams(): DeletePubItemsParams {
     const actionParams: DeletePubItemsParams = {
@@ -27,9 +27,14 @@ export class DeletePubItemsFormComponent {
   }
 
   onSubmit(): void {
-    this.batchSvc.deletePubItems(this.deletePubItemsParams).subscribe(actionResponse => {
-      this.batchSvc.startProcess(actionResponse.batchLogHeaderId);
-      this.router.navigate(['/batch/logs']);
+    let ref = this.msgSvc.displayConfirmation({ text: $localize`:@@batch.actions.item.state.delete.confirmation:Do you really want to delete this items?`, confirm: $localize`:@@confirm:Confirm`, cancel: $localize`:@@cancel:Cancel` });
+    ref.closed.subscribe(confirmed => {
+      if (confirmed) {
+        this.batchSvc.deletePubItems(this.deletePubItemsParams).subscribe(actionResponse => {
+          this.batchSvc.startProcess(actionResponse.batchLogHeaderId);
+          this.router.navigate(['/batch/logs']);
+        });
+      }
     });
   }
 }
